Bet/Result buttons are disabled as per the rules of the game
Now you cannot bet more than the number of hands in the game. The last player to bet can't choose a number that makes the total number of hands bet be equal to the number of hands in the round.
You also can't have gotten more hands than what is left to be had, or less if you are the last player to have results.