New design of the game activity
The table has dividers and the bets show in green (if the result matches the bet) or in red and strike through (if it doesn't). The activity has a bottom sheet which will show the number of hands in the current round, the first player and the dealer, but now it is just a mockup. The FAB is animated when the bottom sheet goes up or down.